remove Autokeys from MyHandler.ts
